+++ src/analyzer.ts
@@ -428,7 +428,12 @@ export class Analyzer {
     }
 
     const files = this.#processFiles(placeholder, packageJson, syntaxInfo);
-    const output = this.#processOutput(placeholder, packageJson, syntaxInfo);
+    const output = this.#processOutput(
+      placeholder,
+      packageJson,
+      syntaxInfo,
+      command
+    );
     const clean = this.#processClean(placeholder, packageJson, syntaxInfo);
     this.#processPackageLocks(placeholder, packageJson, syntaxInfo, files);
 
@@ -654,7 +659,8 @@ export class Analyzer {
   #processOutput(
     placeholder: UnvalidatedConfig,
     packageJson: PackageJson,
-    syntaxInfo: ScriptSyntaxInfo
+    syntaxInfo: ScriptSyntaxInfo,
+    command: JsonAstNode<string> | undefined
   ): undefined | ArrayNode<string> {
     if (syntaxInfo.wireitConfigNode == null) {
       return;
@@ -664,6 +670,25 @@ export class Analyzer {
     ]);
     if (outputNode === undefined) {
       return;
+    }
+    if (command === undefined) {
+      placeholder.failures.push({
+        type: 'failure',
+        reason: 'invalid-config-syntax',
+        script: placeholder,
+        diagnostic: {
+          severity: 'error',
+          message: `"output" can only be set if "command" is also set.`,
+          location: {
+            file: packageJson.jsonFile,
+            range: {
+              // Highlight the whole `"output": []` part.
+              length: (outputNode.parent ?? outputNode).length,
+              offset: (outputNode.parent ?? outputNode).offset,
+            },
+          },
+        },
+      });
     }
     const values = [];
     const result = failUnlessArray(outputNode, packageJson.jsonFile);